% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Displaying year-wise fatality summaries and representing them visually in the form of maps by
state}
\usage{
fars_read(filename)
}
\arguments{
\item{takes}{a year as input,creates a filename, reads the information, adds a column for year then
subsets the information, validates if there is an error and appropriately displays a warning
message if an error occurs else returns "NULL"}

\item{to}{display a monthwise summary of the number of fatal injuries for the year
specified}

\item{To}{display a map plot of the state for the year specified with the dots for each fatal
injury incident. Each state has a state number corresponding to it. The input to this
function is the state number and the year. The data for the state is filtered and the
number of rows is determined. Each fatal injury incident is plotted as dot within the map
at their corresponding latitude and longitude}
}
\value{
If an error is generated a warning message "invalid year: " with
   the invalid year information is displayed. If no error is generated it displays "NULL"

A table displaying a month number and the consolidated number of fatal injuries for
each month of the specified year

This function returns map plot of the state for the year specified with the dots for
   each fatal injury incident with the map at their corresponding latitude and longitude of
   occurence
}
\description{
The dataset used data from the US National Highway Traffic Safety Administration's
    Fatality Analysis Reporting System, which is a nationwide census providing the
    American public yearly data regarding fatal injuries suffered in motor vehicle
    traffic crashes
}
\details{
The function fars_read is used to evaluate the existance of a file and read in the
 information into a dataframe if the file exists

@param to print the message saying the input file name does not exist if the file does not
 @param exist else it reads in csv file that is provided as input. All messages pertaining
   to the progress of the reading in of the csv file is suppresed. A dataframe is created to
   save the input that is read in

@return This function either returns a message the filename does not exist if the file name
    name does not exist or a dataframe containing the information if the file exists



@examples
 fars_read("accident_2013.csv.bz2")
 fars_read("accident_2017.csv.bz2")

The make_filename function is used to create a filename using the year as an input and
 contantenating the text "accident",the year that was inputted and the extensions ".csv.bz2"

@param to print the filename by contantenating the text "accident",the numeric value of the
   year that was inputted and the extensions ".csv.bz2"

@return This function returns a character vector containing the formatted combination of text
    and variable values

@examples
 make_filename(2013)
 make_filename(2014)

The function fars_read_years is used to take the year as an input and create a filename
using make_filename function. The file is then read using the fars_read function. To which a
column called year is added using the mutate function. To use the mutate function the dplyr
package needs to be installed.The data is then subset by month and
year using the dplyr::select function. Trycatch provides a mechanism for handling unusual
conditions. If an error is generated, the error handler picks up the error generated and
displays a warning message "invalid year: " with the invalid year information. If no error is
generated, it displays "NULL". The entire function of creating the files, reading in the
details, adding a year column, subsetting the information is applied to all elements of the
list and returns a list using the lapply function
}
\examples{
fars_read_years(2013)
fars_read_years(2014)


The fars_summarize_years function takes a year as an input and summarizes the number of
accidents by month within the input year specified

fars_summarize_years(2013)
fars_summarize_years(2014)


The fars_map_state function takes the state number and year as an input and displays a plot
of the state with the number of fatalaties mapped as dots at the appropriate location of the
incident. To be able to use this function ensure the maps package is installed

fars_map_state(1,2013)
fars_map_state(4,2013)


}
